Add Controller interface back Move View interface to other
