+++ tst/contoller/ExportControllerTest.java
@@ -12,7 +12,7 @@ import responder.Presenter;
 import responder.Response;
 import usecase.RequestBuilderImpl;
 import usecase.exportfile.ExportRequest;
-import view.View;
+import other.View;
 
 import java.io.File;
 import java.util.HashMap;
